Split campaign info from mission.h/.cpp
